Attractions Page + some css tweaks
Added the attractions page, did the work in all three files to make sure there are little to no contradictions in the code. Also worked on making each page section pit the entire viewport. I thought this would increase usability and aesthetics. I still think we can make some improvements with the appearance, but I'd like to get this code in while it's working well.